feat: add `SessionRole` decorator, enhance `UserController` endpoints with role-based checks, and update service methods for access control

+++ src/user/user.controller.ts
@@ -7,6 +7,7 @@ import {
   ParseIntPipe,
   Patch,
   Post,
+  Req,
   UseGuards,
 } from '@nestjs/common';
 import { UserService } from './user.service';
@@ -17,18 +18,21 @@ import { RoleGuard } from '../auth/guards/role.guard';
 import { RequiredRole } from '../auth/decorators/role.decorator';
 import { UserRole } from './entities/user.enums';
 import { SessionUserId } from '../auth/decorators/session-user.decorator';
+import { SessionRole } from '../auth/decorators/session-role.decorator';
 
-@UseGuards(SessionGuard, RoleGuard)
+@UseGuards(SessionGuard)
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
+  @UseGuards(RoleGuard)
   @RequiredRole(UserRole.ADMIN)
   @Post()
   create(@Body() dto: CreateUserDto) {
     return this.userService.create({ dto });
   }
 
+  @UseGuards(RoleGuard)
   @RequiredRole(UserRole.ADMIN)
   @Get()
   findAll() {
@@ -39,8 +43,10 @@ export class UserController {
   findOne(
     @Param('id', ParseIntPipe) id: number,
     @SessionUserId() userId: number,
+    @SessionRole() role: UserRole,
   ) {
-    return this.userService.findOne({ id, userId });
+    console.log(userId);
+    return this.userService.findOne({ id, userId, role });
   }
 
   @Patch(':id')
@@ -48,15 +54,17 @@ export class UserController {
     @Param('id', ParseIntPipe) id: number,
     @Body() dto: UpdateUserDto,
     @SessionUserId() userId: number,
+    @SessionRole() role: UserRole,
   ) {
-    return this.userService.update({ id, dto, userId });
+    return this.userService.update({ id, dto, userId, role });
   }
 
   @Delete(':id')
   remove(
     @Param('id', ParseIntPipe) id: number,
     @SessionUserId() userId: number,
+    @SessionRole() role: UserRole,
   ) {
-    return this.userService.remove({ id, userId });
+    return this.userService.remove({ id, userId, role });
   }
 }
